// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ImageStatus = {
  OPTIMIZED: 'OPTIMIZED',
  RECOMMENDED: 'RECOMMENDED',
  HIGH_PRIORITY: 'HIGH_PRIORITY',
  FAILED: 'FAILED'
} as const

export type ImageStatus = (typeof ImageStatus)[keyof typeof ImageStatus]


export const JobStatus = {
  QUEUED: 'QUEUED',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus]